use virtual memory instead of dynamic allocation, impl lock free systems & fix multithreaded worker system
we now use the virtual memory of the OS to reserve all the address space we would ever need for an array.
this means pointers for arrays never need to change.
this allowed us to make our arrays lock free and go back to using indexes instead of hold big 8 bytes pointers.